reogarnized RoutesHistory sub-components
